% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformatting_addins.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Wide to long}
\usage{
wide_to_long()
}
\description{
CONVERT: \cr \cr \code{c("bob", "billy", "jeff")} \cr \cr TO: \code{\cr \cr c(\cr"bob",\cr"billy",\cr"jeff"\cr)}
}
